Fix the build, part 1½: Generalize sprite building, fix deps
Before, I was trying very hard to not put any labels in generated code,
which meant I had to include the generated asm into regular source
files, which made for a dependency detection catch-22 that is apparently
somewhat infamous.

I gave up and built sprites to their own source files, which means they
get their own object files, which means the linker just deals with it
all at the end and there's no problem.

Also, this solves the issue of generating a garbage file when something
goes wrong with the Python script.